Work done up to creation of repository
Creating the three main packages: context, value, expression. Everything besides tests will fall under one of these packages. Context will be the logic of the console and what the user mostly interacts with. Values and Expressions are what the context interpret user input as based on the parser in context.